requirements.txt
pyqt5
aiohttp
pyyaml
loguru
